# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Autenticación de usuario",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "admin@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "Registro de usuario",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "profile_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "profile_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "email",
                                        "password",
                                        "profile_id"
                                    ],
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "john@example.com"
                                        },
                                        "profile_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/generateReport": {
            "post": {
                "summary": "Generar informe de juegos",
                "operationId": "8657ecd6ac7b1d2b8f57df787043864e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "report": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 78
                                                },
                                                "player_id": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "1999-12-15"
                                                },
                                                "finish_date": {
                                                    "type": "string",
                                                    "example": "1993-04-05"
                                                },
                                                "points": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_winner": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "amount_envido": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "amount_flower": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-01-24T23:59:56.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-01-24T23:59:56.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 8
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "admin"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "admin@test.com"
                                            },
                                            "profile_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-01-25T00:47:05.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-01-25T00:47:05.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informe generado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "report": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}